+++ plugins/module_utils/module.py
@@ -12,25 +12,36 @@ from ansible.module_utils.six.moves.urllib.error import URLError
 from ..module_utils.base import (
     HVaultClient,
     hvault_argument_spec,
+    hvault_compare,
 )
 
 
-def optspec_to_config(optspec, params, join_lists=False):
+def optspec_to_config(optspec, params):
     config = {}
     for (k, conf) in optspec.items():
         v = params[k]
-        if conf.get('type') == 'list' and join_lists:
-            if v:
-                config[k] = ','.join(v)
-            else:
-                config[k] = ''
-        else:
-            config[k] = v
+        if conf.get('type') == 'list':
+            if conf.get('join'):
+                if v:
+                    v = ','.join(v)
+                else:
+                    v = ''
+        config[k] = v
     return config
 
 
+def optspec_to_argspec(optspec):
+    filtered = {}
+    for (k, v) in optspec.items():
+        dup_v = v.copy()
+        dup_v.pop('join', None)
+        dup_v.pop('unsorted', None)
+        filtered[k] = dup_v
+    return filtered
+
+
 class HVaultModule():
-    def __init__(self, argspec, optspec):
+    def __init__(self, argspec, optspec, **kwargs):
         _argspec = hvault_argument_spec()
         _argspec.update(
             dict(
@@ -46,11 +57,12 @@ class HVaultModule():
         )
         _argspec.update(argspec)
         self.optspec = optspec
-        _argspec.update(optspec)
+        _argspec.update(optspec_to_argspec(optspec))
 
         self.module = AnsibleModule(
             supports_check_mode=True,
             argument_spec=_argspec,
+            **kwargs,
         )
         self.params = self.module.params
         self.client = HVaultClient(self.params, self.module)
@@ -61,7 +73,7 @@ class HVaultModule():
     def mangle_result(self, result):
         return result
 
-    def run(self, path_fmt, config, result_key='role', bad_keys=None, join_lists=False):
+    def run(self, path_fmt, config, result_key='role', bad_keys=None):
         path = path_fmt.format(self.params['mount_point'].strip('/'), self.params['name'])
         self._path = path
 
@@ -81,10 +93,14 @@ class HVaultModule():
                     self.client.delete(path)
             self.module.exit_json(changed=changed)
 
-        config.update(optspec_to_config(self.optspec, self.params, join_lists))
+        config.update(optspec_to_config(self.optspec, self.params))
         config = self.mangle_config(config)
 
-        if config != result:
+        if not hvault_compare(
+            config, result,
+            ignore_keys=bad_keys,
+            unsorted_keys=[k for (k, v) in self.optspec.items() if v.get('unsorted')],
+        ):
             changed = True
             if not self.module.check_mode:
                 self.client.post(path, config)